+++ js/main.js
@@ -1,67 +1,5 @@
-
-const COMMENT_MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const NAMES = [
-  'Иван',
-  'Хуан Себастьян',
-  'Мария',
-  'Кристоф',
-  'Виктор',
-  'Юлия',
-  'Люпита',
-  'Вашингтон',
-];
-
-/**
- * Получает случайное число из диапазона
- * @param {number} a - Минимальное значение диапазона случайных чисел
- * @param {number} b - Максимальное значение диапазона случайных чисел
- * @returns {int} result - Случайное число из диапазона
- */
-function getRandomInteger (a, b) {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-}
-
-/**
- * Получает случайное уникальное число из указанного диапазона
- * @param {number} min - Минимальное значение диапазона случайных чисел
- * @param {number} max - Максимальное значение диапазона случайных чисел
- * @returns {int} newUniqueNumber - Случайное уникальное число из диапазона
- */
-
-function getRandomUniqueNumberFromRange (min = 1, max = 25) {
-  const previousUniqueNumbersArray = [];
-
-  return function () {
-    let newUniqueNumber = getRandomInteger(min, max);
-    if (previousUniqueNumbersArray.length >= (max - min + 1)) {
-      return null;
-    }
-    while (previousUniqueNumbersArray.includes(newUniqueNumber)) {
-      newUniqueNumber = getRandomInteger(min, max);
-    }
-    previousUniqueNumbersArray.push(newUniqueNumber);
-    return newUniqueNumber;
-  };
-}
-
-/**
- * @param {Array} element - Массив, случайный элемент которого необходимо получить
- * @returns {any}
- */
-function getRandomElementFromArray (element) {
-  return element[getRandomInteger(0, element.length - 1)];
-}
+import {COMMENT_MESSAGES, NAMES} from '/data.js';
+import {getRandomInteger, getRandomUniqueNumberFromRange, getRandomElementFromArray} from '/util.js';
 
 const photoID = getRandomUniqueNumberFromRange();
 const imageNameNumber = getRandomUniqueNumberFromRange();
@@ -87,10 +25,10 @@ function getGeneratedPhoto () {
   };
 }
 
-function generatePhotos () {
-  return Array.from({length: 25}, getGeneratedPhoto);
+function generatePhotos (length) {
+  return Array.from({length: length}, getGeneratedPhoto);
 }
 
-const randomObjectsArray = generatePhotos();
+const randomObjectsArray = generatePhotos(25);
 
 console.table(randomObjectsArray);
